Update crate to use Tokio reform internally (#303)
Tokio reform was released as a separate crate (`tokio`). However, the
ecosystem has been split as tokio-core is not compatible with tokio.
Libraries (like hyper) that expose tokio-core as part of the public API
cannot easily switch to tokio as is.

This patch updates tokio-core to use tokio internally. This should help
reduce the friction encountered when users attempt to update their code
bases to tokio.